Refactor registration page to use server-side action and add form reset functionality
